requirements.txt
openai>=1.57.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
textblob>=0.17.0
newspaper3k>=0.2.8
feedparser>=6.0.0
